Add commitlog to MERGE_MSG files
These files are used for preparing the commit message when we are
merging another branch.
It's pretty new as it used to not be used when the merge was
successful.

Now even if the merge is successful, you are given the chance to
explain why you merged in the feature.

Let's use the commit-log mode to these files as the same rules apply.