Add Uniquify as an optional dependency
It's pretty annoying when opening multiple files with the same name that
they are uniquely named by appending a meaning-less number. That
situation happens a lot when modifing build configuration files
(Makefile, SConscript, ...) as they are usually all named the same in
different directories.

Use the Uniquify module for that with a decent enough configuration so
that the confusion is reduced as much as possible.